Convert 'source_str' from list of str to ndarray
Input data batches from TextImageGenerator contain `source_str`, which
is used inside the visualization callback. It is a list of strings that
correspond to the word represented by each image. Tensorflow throws an
error because it expects a Tensor, but `source_str` can't be
converted to a Tensor. Using np.array(`source_str`) fixes this.