Remove compatibility for 'channels_first'
In multi-backend Keras, channel ordering can be set to 'channels_first'
to provide support for Theano. The old tutorial was written for
multi-backend Keras, so it checked which mode Keras was configured for.
Compatibility for both formats is no longer necessary, because tf.keras
is written only for Tensorflow (which uses 'channels_last'.)